+++ src/snowflake/cli/_plugins/notebook/notebook_entity.py
@@ -4,9 +4,8 @@ from click import ClickException
 from snowflake.cli._plugins.connection.util import make_snowsight_url
 from snowflake.cli._plugins.notebook.notebook_entity_model import NotebookEntityModel
 from snowflake.cli._plugins.notebook.notebook_project_paths import NotebookProjectPaths
-from snowflake.cli._plugins.stage.manager import StageManager
 from snowflake.cli._plugins.workspace.context import ActionContext
-from snowflake.cli.api.artifacts.utils import bundle_artifacts
+from snowflake.cli.api.artifacts.upload import sync_artifacts_with_stage
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.console.console import cli_console
 from snowflake.cli.api.entities.common import EntityBase
@@ -22,12 +21,15 @@ class NotebookEntity(EntityBase[NotebookEntityModel]):
     A notebook.
     """
 
+    @property
+    def _stage_path_from_model(self) -> str:
+        if self.model.stage_path is None:
+            return f"{_DEFAULT_NOTEBOOK_STAGE_NAME}/{self.fqn.name}"
+        return self.model.stage_path
+
     @functools.cached_property
     def _stage_path(self) -> StagePath:
-        stage_path = self.model.stage_path
-        if stage_path is None:
-            stage_path = f"{_DEFAULT_NOTEBOOK_STAGE_NAME}/{self.fqn.name}"
-        return StagePath.from_stage_str(stage_path)
+        return StagePath.from_stage_str(self._stage_path_from_model)
 
     @functools.cached_property
     def _project_paths(self):
@@ -40,26 +42,6 @@ class NotebookEntity(EntityBase[NotebookEntityModel]):
             return True
         except ProgrammingError:
             return False
-
-    def _upload_artifacts(self):
-        stage_fqn = self._stage_path.stage_fqn
-        stage_manager = StageManager()
-        cli_console.step(f"Creating stage {stage_fqn} if not exists")
-        stage_manager.create(fqn=stage_fqn)
-
-        cli_console.step("Uploading artifacts")
-
-        # creating bundle map to handle glob patterns logic
-        bundle_map = bundle_artifacts(self._project_paths, self.model.artifacts)
-        for absolute_src, absolute_dest in bundle_map.all_mappings(
-            absolute=True, expand_directories=True
-        ):
-            artifact_stage_path = self._stage_path / (
-                absolute_dest.relative_to(self._project_paths.bundle_root).parent
-            )
-            stage_manager.put(
-                local_path=absolute_src, stage_path=artifact_stage_path, overwrite=True
-            )
 
     def get_create_sql(self, replace: bool) -> str:
         main_file_stage_path = self._stage_path / (
@@ -99,6 +81,7 @@ class NotebookEntity(EntityBase[NotebookEntityModel]):
         self,
         action_ctx: ActionContext,
         replace: bool,
+        prune: bool,
         *args,
         **kwargs,
     ) -> str:
@@ -108,7 +91,13 @@ class NotebookEntity(EntityBase[NotebookEntityModel]):
                     f"Notebook {self.fqn.name} already exists. Consider using --replace."
                 )
         with cli_console.phase(f"Uploading artifacts to {self._stage_path}"):
-            self._upload_artifacts()
+            sync_artifacts_with_stage(
+                project_paths=self._project_paths,
+                stage_root=self._stage_path_from_model,
+                prune=prune,
+                artifacts=self.model.artifacts,
+            )
+
         with cli_console.phase(f"Creating notebook {self.fqn}"):
             return self.action_create(replace=replace)