+++ tests/test_connection.py
@@ -1546,3 +1546,83 @@ def test_connection_add_no_key_pair_setup_if_no_interactive(
     assert (
         result.output.strip() == f"Wrote new connection conn to {test_snowcli_config}"
     )
+
+
+@mock.patch("snowflake.cli._plugins.auth.keypair.manager.AuthManager.execute_query")
+@mock.patch("snowflake.cli._plugins.object.manager.ObjectManager.execute_query")
+@mock.patch("snowflake.connector.connect")
+def test_connection_add_with_key_pair_saves_password_if_keypair_is_set(
+    mock_connect,
+    mock_object_execute_query,
+    mock_auth_execute_query,
+    runner,
+    tmp_path,
+    mock_cursor,
+    test_snowcli_config,
+):
+    mock_connect.return_value.user = "user"
+    mock_object_execute_query.return_value = mock_cursor(
+        rows=[
+            {"property": "RSA_PUBLIC_KEY", "value": None},
+            {"property": "RSA_PUBLIC_KEY_2", "value": "public key..."},
+        ],
+        columns=[],
+    )
+
+    result = runner.invoke(
+        [
+            "connection",
+            "add",
+        ],
+        input="conn\n"  # connection name: zz
+        "test\n"  # account:
+        "user\n"  # user:
+        "123\n"  # password:
+        "\n"  # role:
+        "\n"  # warehouse:
+        "\n"  # database:
+        "\n"  # schema:
+        "\n"  # host:
+        "\n"  # port:
+        "\n"  # region:
+        "\n"  # authenticator:
+        "\n"  # private key file:
+        "\n"  # token file path:
+        "y\n"  #
+        "\n"  # key_length
+        f"{tmp_path}\n"  # output_path
+        "123\n",  # passphrase
+    )
+
+    private_key_path = tmp_path / "conn.p8"
+    public_key_path = tmp_path / "conn.pub"
+    assert result.exit_code == 0, result.output
+    assert result.output == dedent(
+        f"""\
+        Enter connection name: conn
+        Enter account: test
+        Enter user: user
+        Enter password: 
+        Enter role: 
+        Enter warehouse: 
+        Enter database: 
+        Enter schema: 
+        Enter host: 
+        Enter port: 
+        Enter region: 
+        Enter authenticator: 
+        Enter private key file: 
+        Enter token file path: 
+        Do you want to configure key pair authentication? [y/N]: y
+        Key length [2048]: 
+        Output path [~/.ssh]: {tmp_path}
+        Private key passphrase: 
+        Wrote new password-based connection conn to {test_snowcli_config}, however there were some issues during key pair setup. Review the following error and check 'snow auth keypair' commands to setup key pair authentication:
+         * The public key is set already.
+        """
+    )
+    assert not private_key_path.exists()
+    assert not public_key_path.exists()
+    with open(test_snowcli_config, "r") as f:
+        connections = tomlkit.load(f)
+        assert connections["connections"]["conn"]["password"] == "123"